refactor for new structure
