+++ pkg/operator/operatorclient/interfaces.go
@@ -1,0 +1,9 @@
+package operatorclient
+
+const (
+	GlobalUserSpecifiedConfigNamespace    = "openshift-config"
+	MachineSpecifiedGlobalConfigNamespace = "openshift-config-managed"
+	OperatorNamespace                     = "openshift-kube-controller-manager-operator"
+	TargetNamespace                       = "openshift-kube-controller-manager"
+	ServiceCertSignerNamespace            = "openshift-service-cert-signer"
+)
