--- pkg/operator/target_config_reconciler_v311_00.go
@@ -1,107 +1,0 @@
-package operator
-
-import (
-	"fmt"
-	"reflect"
-
-	corev1 "k8s.io/api/core/v1"
-	coreclientv1 "k8s.io/client-go/kubernetes/typed/core/v1"
-
-	operatorv1 "github.com/openshift/api/operator/v1"
-	"github.com/openshift/library-go/pkg/operator/events"
-	"github.com/openshift/library-go/pkg/operator/resource/resourceapply"
-	"github.com/openshift/library-go/pkg/operator/resource/resourcemerge"
-	"github.com/openshift/library-go/pkg/operator/resource/resourceread"
-	"github.com/openshift/library-go/pkg/operator/v1helpers"
-
-	"github.com/openshift/cluster-kube-controller-manager-operator/pkg/apis/kubecontrollermanager/v1alpha1"
-	"github.com/openshift/cluster-kube-controller-manager-operator/pkg/operator/v311_00_assets"
-	"github.com/openshift/cluster-kube-controller-manager-operator/pkg/version"
-)
-
-// syncKubeControllerManager_v311_00_to_latest takes care of synchronizing (not upgrading) the thing we're managing.
-// most of the time the sync method will be good for a large span of minor versions
-func createTargetConfigReconciler_v311_00_to_latest(c TargetConfigReconciler, recorder events.Recorder, operatorConfig *v1alpha1.KubeControllerManagerOperatorConfig) (bool, error) {
-	operatorConfigOriginal := operatorConfig.DeepCopy()
-	errors := []error{}
-
-	directResourceResults := resourceapply.ApplyDirectly(c.kubeClient, c.eventRecorder, v311_00_assets.Asset,
-		"v3.11.0/kube-controller-manager/ns.yaml",
-		"v3.11.0/kube-controller-manager/svc.yaml",
-		"v3.11.0/kube-controller-manager/sa.yaml",
-	)
-	for _, currResult := range directResourceResults {
-		if currResult.Error != nil {
-			errors = append(errors, fmt.Errorf("%q (%T): %v", currResult.File, currResult.Type, currResult.Error))
-		}
-	}
-
-	_, _, err := manageKubeControllerManagerConfigMap_v311_00_to_latest(c.kubeClient.CoreV1(), recorder, operatorConfig)
-	if err != nil {
-		errors = append(errors, fmt.Errorf("%q: %v", "configmap", err))
-	}
-	_, _, err = resourceapply.SyncConfigMap(c.kubeClient.CoreV1(), recorder, serviceCertSignerNamespaceName, "signing-cabundle", targetNamespaceName, "signing-cabundle", nil)
-	if err != nil {
-		errors = append(errors, fmt.Errorf("%q: %v", "configmap", err))
-	}
-	_, _, err = managePod_v311_00_to_latest(c.kubeClient.CoreV1(), recorder, operatorConfig, c.targetImagePullSpec)
-	if err != nil {
-		errors = append(errors, fmt.Errorf("%q: %v", "configmap/kube-controller-manager-pod", err))
-	}
-
-	if len(errors) > 0 {
-		message := ""
-		for _, err := range errors {
-			message = message + err.Error() + "\n"
-		}
-		v1helpers.SetOperatorCondition(&operatorConfig.Status.Conditions, operatorv1.OperatorCondition{
-			Type:    "TargetConfigReconcilerFailing",
-			Status:  operatorv1.ConditionTrue,
-			Reason:  "SynchronizationError",
-			Message: message,
-		})
-		if !reflect.DeepEqual(operatorConfigOriginal, operatorConfig) {
-			_, updateError := c.operatorConfigClient.KubeControllerManagerOperatorConfigs().UpdateStatus(operatorConfig)
-			return true, updateError
-		}
-		return true, nil
-	}
-
-	v1helpers.SetOperatorCondition(&operatorConfig.Status.Conditions, operatorv1.OperatorCondition{
-		Type:   "TargetConfigReconcilerFailing",
-		Status: operatorv1.ConditionFalse,
-	})
-	if !reflect.DeepEqual(operatorConfigOriginal, operatorConfig) {
-		_, updateError := c.operatorConfigClient.KubeControllerManagerOperatorConfigs().UpdateStatus(operatorConfig)
-		if updateError != nil {
-			return true, updateError
-		}
-	}
-
-	return false, nil
-}
-
-func manageKubeControllerManagerConfigMap_v311_00_to_latest(client coreclientv1.ConfigMapsGetter, recorder events.Recorder, operatorConfig *v1alpha1.KubeControllerManagerOperatorConfig) (*corev1.ConfigMap, bool, error) {
-	configMap := resourceread.ReadConfigMapV1OrDie(v311_00_assets.MustAsset("v3.11.0/kube-controller-manager/cm.yaml"))
-	defaultConfig := v311_00_assets.MustAsset("v3.11.0/kube-controller-manager/defaultconfig.yaml")
-	requiredConfigMap, _, err := resourcemerge.MergeConfigMap(configMap, "config.yaml", nil, defaultConfig, operatorConfig.Spec.ObservedConfig.Raw, operatorConfig.Spec.UnsupportedConfigOverrides.Raw)
-	if err != nil {
-		return nil, false, err
-	}
-	return resourceapply.ApplyConfigMap(client, recorder, requiredConfigMap)
-}
-
-func managePod_v311_00_to_latest(client coreclientv1.ConfigMapsGetter, recorder events.Recorder, operatorConfig *v1alpha1.KubeControllerManagerOperatorConfig, imagePullSpec string) (*corev1.ConfigMap, bool, error) {
-	required := resourceread.ReadPodV1OrDie(v311_00_assets.MustAsset("v3.11.0/kube-controller-manager/pod.yaml"))
-	required.Spec.Containers[0].ImagePullPolicy = corev1.PullAlways
-	if len(imagePullSpec) > 0 {
-		required.Spec.Containers[0].Image = imagePullSpec
-	}
-	required.Spec.Containers[0].Args = append(required.Spec.Containers[0].Args, fmt.Sprintf("-v=%d", 4))
-
-	configMap := resourceread.ReadConfigMapV1OrDie(v311_00_assets.MustAsset("v3.11.0/kube-controller-manager/pod-cm.yaml"))
-	configMap.Data["pod.yaml"] = resourceread.WritePodV1OrDie(required)
-	configMap.Data["forceRedeploymentReason"] = operatorConfig.Spec.ForceRedeploymentReason
-	configMap.Data["version"] = version.Get().String()
-	return resourceapply.ApplyConfigMap(client, recorder, configMap)
-}
